Oops, fixed type-o's in member variable name and protection state.
I only unit tested the core part of the script (the actual Star English
Generator) but didn't test the file IO part (open source file, save
original file, etc.) and turns out when I converted the script into a
class and made the source_translation_file variable a member variable
(vs. a input parameter) I missed some as the file IO methods called it
source_translation_filename (filename vs. file). This change set updates
the script to use source_translation_filename as the variable name.

Also the method I did test I originally had private but then changed to
protected but failed to update the entry method so it was still trying
to access a private version of the method which does not exist. The unit
tests passed as I bypassed the entry method and tested the protected
method directly. This change set updates the entry method to access the
right do'er method.