Adding unit tests for the Python version of the Star English generator
this change set also converts the Python script into a class so that we
can snub out the file IO calls for the unit tests. It also includes two
new unit tests (which currently fail) that outlines other forms of Qts
TS XML that one could run into that this script needs to support (but
currently doesn't).