Adding support for the plural form of qsTr()/tr() to the Python version.
This change set adds support for the numerus tag which is used when you
use the plural form of qsTr() or tr()  - i.e. qsTr( "Month(s)", "", n)

Note the script will take the original string and clip off the "(s)"
when generating Star-English for the singular form or remove the
parentheses for the plural form. i.e. Month(s) becomes *Month* and
*Months*