Adding a Long-English version of the Python script
Refactored the Star-English Python script so that I can create a
Long-English version simply by sub-classing a base class. Now both
Star-English and Long-English derive from the same base class
(TranslationGenerator) as they both share most of the same code.

Long-English works the same way Star-English does but instead of
wrapping the English string in * it doubles the length of the string by
wrapping the string in W's.