Implemented changes to make tests pass.
I had created a number of unit tests which outlined pending changes I
wanted to make but had yet to make. This change set implements those
changes to make the unit tests pass. Now the generator will correctly
handle the introduction of a number of comment tags, will provide
singular and plural translations if required, change the language to the
given Star-English language in the root tag (i..e fo_FO), and will error
if the XML read is not what we're expecting (i.e. is valid XML but not
the right XML as far as the generator cares about).

This change set also implements a new menu in the generator UI to allow
the user to pick different languages to show the tool in. Currently the
only supported languages are English and Star-English.