Added check  so as to not blow away the source file
Currently if you set the input and output files to be the same the tool
will open the source in read-only mode then open the destination (which
in this case is the same file as the source) in write-only mode which
blanks out the file (because we are not opening in Append mode). The
generation then fails because the source file is empty. This runs the
risk of users accedently erasing the users source file.

This change set adds a check to make sure that we don't try to open the
same IO device in read-only and write-only modes preventing the erasing
of the source data.

As the generator works on QIODevices which are not able to tell if  two
QOIDevices are pointing to the same source the check is put into the
BaseGenerationCommand class which works on QFiles and does a file name
comparison.